% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acs_pums_sample}
\alias{acs_pums_sample}
\title{ACS PUMS-like sample}
\format{
A data frame with 500 rows and 5 variables:
\describe{
  \item{age}{age in years}
  \item{income}{annual income in USD}
  \item{educ}{educational attainment}
  \item{married}{married indicator}
  \item{weeks_worked}{weeks worked last year}
}
}
\usage{
acs_pums_sample
}
\description{
A synthetic microdata sample modeled on American Community Survey PUMS
structure, used as an example in \pkg{DPSynth}.
}
\keyword{datasets}
