% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounting.R
\name{new_synth_budget}
\alias{new_synth_budget}
\title{Create a privacy budget tracker}
\usage{
new_synth_budget(
  epsilon,
  delta = 0,
  accounting = c("basic", "advanced", "rdp")
)
}
\arguments{
\item{epsilon}{total epsilon available}

\item{delta}{total delta available}

\item{accounting}{one of \code{"basic"}, \code{"advanced"}, \code{"rdp"}}
}
\value{
object of class \code{synth_privacy_budget}
}
\description{
Create a privacy budget tracker
}
