% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesis.R
\name{print.dp_synthesis_result}
\alias{print.dp_synthesis_result}
\title{Print a synthesis result}
\usage{
\method{print}{dp_synthesis_result}(x, ...)
}
\arguments{
\item{x}{a \code{dp_synthesis_result}}

\item{...}{unused}
}
\description{
Print a synthesis result
}
